refactor(cache): replace CacheManager stubs with handler-only narrowing #1230
Params defer to reflection for real manager calls, so Laravel's 13.5.0
UnitEnum widening needs no duplicated version-dir signature; the facade
pseudo-method path keeps copying the canonical @method params. Both
providers share one real-vs-pseudo discriminator (methodExists without
pseudo) so the narrowing also covers CacheManager subclasses.
