--- stubs/13.5.0/Cache/CacheManager.phpstub
@@ -1,34 +1,0 @@
-<?php
-
-namespace Illuminate\Cache;
-
-use Illuminate\Contracts\Cache\Factory as FactoryContract;
-
-/**
- * Laravel 13.5.0 widened the store()/driver()/memo() name parameter to \UnitEnum;
- * see stubs/common/Cache/CacheManager.phpstub for the narrowing rationale (#1230)
- * and the verbatim-header requirement.
- *
- * @mixin \Illuminate\Cache\Repository
- * @mixin \Illuminate\Contracts\Cache\LockProvider
- */
-class CacheManager implements FactoryContract
-{
-    /**
-     * @param  \UnitEnum|string|null  $name
-     * @return \Illuminate\Cache\Repository
-     */
-    public function store($name = null) {}
-
-    /**
-     * @param  \UnitEnum|string|null  $driver
-     * @return \Illuminate\Cache\Repository
-     */
-    public function driver($driver = null) {}
-
-    /**
-     * @param  \UnitEnum|string|null  $driver
-     * @return \Illuminate\Cache\Repository
-     */
-    public function memo($driver = null) {}
-}
