--- stubs/common/Cache/CacheManager.phpstub
@@ -1,39 +1,0 @@
-<?php
-
-namespace Illuminate\Cache;
-
-use Illuminate\Contracts\Cache\Factory as FactoryContract;
-
-/**
- * store()/driver()/memo() document the Repository CONTRACT, but at runtime the
- * value is the concrete \Illuminate\Cache\Repository: built-in drivers are wrapped
- * by repository(), memo() wraps in a MemoizedStore, and Cache::extend() creators
- * return Cache::repository(...) by documented convention (not enforced). Narrowing
- * restores the concrete-only surface: flexible(), tags(), Macroable. See #1230.
- *
- * Class declaration mirrors Laravel's source: a stub re-declaration wipes the
- * reflected class_implements list, so the clause must be repeated verbatim.
- *
- * @mixin \Illuminate\Cache\Repository
- * @mixin \Illuminate\Contracts\Cache\LockProvider
- */
-class CacheManager implements FactoryContract
-{
-    /**
-     * @param  string|null  $name
-     * @return \Illuminate\Cache\Repository
-     */
-    public function store($name = null) {}
-
-    /**
-     * @param  string|null  $driver
-     * @return \Illuminate\Cache\Repository
-     */
-    public function driver($driver = null) {}
-
-    /**
-     * @param  string|null  $driver
-     * @return \Illuminate\Cache\Repository
-     */
-    public function memo($driver = null) {}
-}
