+++ tests/Type/tests/Cache/CacheManagerConcreteRepositoryTest.phpt
@@ -4,13 +4,17 @@
 namespace App;
 
 use Illuminate\Cache\CacheManager;
-use Illuminate\Cache\Repository;
 use Illuminate\Support\Facades\Cache;
 
 /**
- * store()/driver()/memo() are narrowed to the concrete Repository, so
- * concrete-only methods like flexible() resolve without errors (#1230).
+ * store()/driver()/memo() declare the Repository interface but always return the
+ * concrete \Illuminate\Cache\Repository, so its concrete-only surface (flexible(),
+ * tags(), Macroable helpers) must resolve on every access path. See issue #1230.
  */
+class CustomCacheManager extends CacheManager
+{
+}
+
 function issue_repro(): array
 {
     return Cache::driver('file')->flexible('things', [300, 3600], static function (): array {
@@ -18,7 +22,7 @@ function issue_repro(): array
     });
 }
 
-function facade_returns_concrete(): void
+function facade(): void
 {
     $_store = Cache::store();
     /** @psalm-check-type-exact $_store = \Illuminate\Cache\Repository */
@@ -30,22 +34,28 @@ function facade_returns_concrete(): void
     /** @psalm-check-type-exact $_memo = \Illuminate\Cache\Repository */
 }
 
-function manager_returns_concrete(CacheManager $manager): void
+function alias(): void
+{
+    $_driver = \Cache::driver();
+    /** @psalm-check-type-exact $_driver = \Illuminate\Cache\Repository */
+}
+
+function injected(CacheManager $manager): void
 {
     $_store = $manager->store();
     /** @psalm-check-type-exact $_store = \Illuminate\Cache\Repository */
-
-    $_driver = $manager->driver();
-    /** @psalm-check-type-exact $_driver = \Illuminate\Cache\Repository */
-
-    $_memo = $manager->memo();
-    /** @psalm-check-type-exact $_memo = \Illuminate\Cache\Repository */
 
     $_flexible = $manager->driver()->flexible('things', [300, 3600], static fn (): int => 1);
     /** @psalm-check-type-exact $_flexible = 1 */
 }
 
-function helper_returns_concrete(): void
+function subclass(CustomCacheManager $manager): void
+{
+    $_driver = $manager->driver();
+    /** @psalm-check-type-exact $_driver = \Illuminate\Cache\Repository */
+}
+
+function helper(): void
 {
     $_driver = cache()->driver();
     /** @psalm-check-type-exact $_driver = \Illuminate\Cache\Repository */
